+++ src/Graph/Nodes/Ball.php
@@ -5,7 +5,7 @@ namespace Rubix\ML\Graph\Nodes;
 use Rubix\ML\Helpers\Stats;
 use Rubix\ML\Datasets\Labeled;
 use Rubix\ML\Kernels\Distance\Distance;
-use Traversable;
+use Rubix\ML\Graph\Nodes\Traits\HasBinaryChildrenTrait;
 
 use function Rubix\ML\argmax;
 
@@ -22,6 +22,8 @@ use function Rubix\ML\argmax;
  */
 class Ball implements Hypersphere, BinaryNode, HasBinaryChildren
 {
+    use HasBinaryChildrenTrait;
+
     /**
      * The center or multivariate mean of the centroid.
      *
@@ -42,20 +44,6 @@ class Ball implements Hypersphere, BinaryNode, HasBinaryChildren
      * @var list<\Rubix\ML\Datasets\Labeled>
      */
     protected array $groups;
-
-    /**
-     * The left child node.
-     *
-     * @var \Rubix\ML\Graph\Nodes\BinaryNode|null
-     */
-    protected ?\Rubix\ML\Graph\Nodes\BinaryNode $left = null;
-
-    /**
-     * The right child node.
-     *
-     * @var \Rubix\ML\Graph\Nodes\BinaryNode|null
-     */
-    protected ?\Rubix\ML\Graph\Nodes\BinaryNode $right = null;
 
     /**
      * Factory method to build a hypersphere by splitting the dataset into left and right clusters.
@@ -149,83 +137,6 @@ class Ball implements Hypersphere, BinaryNode, HasBinaryChildren
     public function groups() : array
     {
         return $this->groups;
-    }
-
-    /**
-     * Return the left child node.
-     *
-     * @return \Rubix\ML\Graph\Nodes\BinaryNode|null
-     */
-    public function left() : ?BinaryNode
-    {
-        return $this->left;
-    }
-
-    /**
-     * Return the right child node.
-     *
-     * @return \Rubix\ML\Graph\Nodes\BinaryNode|null
-     */
-    public function right() : ?BinaryNode
-    {
-        return $this->right;
-    }
-
-    /**
-     * Return the children of this node in a generator.
-     *
-     * @return \Generator<\Rubix\ML\Graph\Nodes\BinaryNode>
-     */
-    public function children() : Traversable
-    {
-        if ($this->left) {
-            yield $this->left;
-        }
-
-        if ($this->right) {
-            yield $this->right;
-        }
-    }
-
-    /**
-     * Return the height of the node in the tree.
-     *
-     * @return int
-     */
-    public function height() : int
-    {
-        return 1 + max($this->left ? $this->left->height() : 0, $this->right ? $this->right->height() : 0);
-    }
-
-    /**
-     * The balance factor of the node. Negative numbers indicate a lean to the left, positive
-     * to the right, and 0 is perfectly balanced.
-     *
-     * @return int
-     */
-    public function balance() : int
-    {
-        return ($this->right ? $this->right->height() : 0) - ($this->left ? $this->left->height() : 0);
-    }
-
-    /**
-     * Set the left child node.
-     *
-     * @param \Rubix\ML\Graph\Nodes\BinaryNode $node
-     */
-    public function attachLeft(?BinaryNode $node = null) : void
-    {
-        $this->left = $node;
-    }
-
-    /**
-     * Set the right child node.
-     *
-     * @param \Rubix\ML\Graph\Nodes\BinaryNode $node
-     */
-    public function attachRight(?BinaryNode $node = null) : void
-    {
-        $this->right = $node;
     }
 
     /**
