+++ src/Graph/Nodes/Box.php
@@ -3,6 +3,7 @@
 namespace Rubix\ML\Graph\Nodes;
 
 use Rubix\ML\Datasets\Labeled;
+use Rubix\ML\Graph\Nodes\Traits\HasBinaryChildrenTrait;
 use Traversable;
 
 use function Rubix\ML\argmax;
@@ -21,6 +22,8 @@ use function Rubix\ML\argmax;
  */
 class Box implements Hypercube, BinaryNode, HasBinaryChildren
 {
+    use HasBinaryChildrenTrait;
+
     /**
      * The feature column (index) of the split value.
      *
@@ -55,20 +58,6 @@ class Box implements Hypercube, BinaryNode, HasBinaryChildren
      * @var list<int|float>
      */
     protected array $max;
-
-    /**
-     * The left child node.
-     *
-     * @var \Rubix\ML\Graph\Nodes\BinaryNode|null
-     */
-    protected ?\Rubix\ML\Graph\Nodes\BinaryNode $left = null;
-
-    /**
-     * The right child node.
-     *
-     * @var \Rubix\ML\Graph\Nodes\BinaryNode|null
-     */
-    protected ?\Rubix\ML\Graph\Nodes\BinaryNode $right = null;
 
     /**
      * Factory method to build a coordinate node from a labeled dataset
@@ -162,83 +151,6 @@ class Box implements Hypercube, BinaryNode, HasBinaryChildren
     public function isPoint() : bool
     {
         return $this->min == $this->max;
-    }
-
-    /**
-     * Return the left child node.
-     *
-     * @return \Rubix\ML\Graph\Nodes\BinaryNode|null
-     */
-    public function left() : ?BinaryNode
-    {
-        return $this->left;
-    }
-
-    /**
-     * Return the right child node.
-     *
-     * @return \Rubix\ML\Graph\Nodes\BinaryNode|null
-     */
-    public function right() : ?BinaryNode
-    {
-        return $this->right;
-    }
-
-    /**
-     * Return the children of this node in a generator.
-     *
-     * @return \Generator<\Rubix\ML\Graph\Nodes\BinaryNode>
-     */
-    public function children() : Traversable
-    {
-        if ($this->left) {
-            yield $this->left;
-        }
-
-        if ($this->right) {
-            yield $this->right;
-        }
-    }
-
-    /**
-     * Recursive function to determine the height of the node in the tree.
-     *
-     * @return int
-     */
-    public function height() : int
-    {
-        return 1 + max($this->left ? $this->left->height() : 0, $this->right ? $this->right->height() : 0);
-    }
-
-    /**
-     * The balance factor of the node. Negative numbers indicate a lean to the left, positive
-     * to the right, and 0 is perfectly balanced.
-     *
-     * @return int
-     */
-    public function balance() : int
-    {
-        return ($this->right ? $this->right->height() : 0) - ($this->left ? $this->left->height() : 0);
-    }
-
-    /**
-     * Set the left child node.
-     *
-     * @param \Rubix\ML\Graph\Nodes\BinaryNode $node
-     */
-    public function attachLeft(?BinaryNode $node = null) : void
-    {
-        $this->left = $node;
-    }
-
-    /**
-     * Set the right child node.
-     *
-     * @param \Rubix\ML\Graph\Nodes\BinaryNode $node
-     */
-    public function attachRight(?BinaryNode $node = null) : void
-    {
-        $this->right = $node;
     }
 
     /**
