+++ src/Graph/Nodes/Isolator.php
@@ -3,7 +3,7 @@
 namespace Rubix\ML\Graph\Nodes;
 
 use Rubix\ML\Datasets\Dataset;
-use Traversable;
+use Rubix\ML\Graph\Nodes\Traits\HasBinaryChildrenTrait;
 
 use function array_unique;
 use function array_rand;
@@ -28,6 +28,8 @@ use function rand;
  */
 class Isolator implements BinaryNode, HasBinaryChildren
 {
+    use HasBinaryChildrenTrait;
+
     /**
      * The feature column (index) of the split value.
      *
@@ -48,20 +50,6 @@ class Isolator implements BinaryNode, HasBinaryChildren
      * @var list<\Rubix\ML\Datasets\Dataset>
      */
     protected array $groups;
-
-    /**
-     * The left child node.
-     *
-     * @var \Rubix\ML\Graph\Nodes\BinaryNode|null
-     */
-    protected ?\Rubix\ML\Graph\Nodes\BinaryNode $left = null;
-
-    /**
-     * The right child node.
-     *
-     * @var \Rubix\ML\Graph\Nodes\BinaryNode|null
-     */
-    protected ?\Rubix\ML\Graph\Nodes\BinaryNode $right = null;
 
     /**
      * Factory method to build a isolator node from a dataset using a random split of the dataset.
@@ -139,83 +127,6 @@ class Isolator implements BinaryNode, HasBinaryChildren
     public function groups() : array
     {
         return $this->groups;
-    }
-
-    /**
-     * Return the left child node.
-     *
-     * @return \Rubix\ML\Graph\Nodes\BinaryNode|null
-     */
-    public function left() : ?BinaryNode
-    {
-        return $this->left;
-    }
-
-    /**
-     * Return the right child node.
-     *
-     * @return \Rubix\ML\Graph\Nodes\BinaryNode|null
-     */
-    public function right() : ?BinaryNode
-    {
-        return $this->right;
-    }
-
-    /**
-     * Return the children of this node in a generator.
-     *
-     * @return \Generator<\Rubix\ML\Graph\Nodes\BinaryNode>
-     */
-    public function children() : Traversable
-    {
-        if ($this->left) {
-            yield $this->left;
-        }
-
-        if ($this->right) {
-            yield $this->right;
-        }
-    }
-
-    /**
-     * Recursive function to determine the height of the node in the tree.
-     *
-     * @return int
-     */
-    public function height() : int
-    {
-        return 1 + max($this->left ? $this->left->height() : 0, $this->right ? $this->right->height() : 0);
-    }
-
-    /**
-     * The balance factor of the node. Negative numbers indicate a lean to the left, positive
-     * to the right, and 0 is perfectly balanced.
-     *
-     * @return int
-     */
-    public function balance() : int
-    {
-        return ($this->right ? $this->right->height() : 0) - ($this->left ? $this->left->height() : 0);
-    }
-
-    /**
-     * Set the left child node.
-     *
-     * @param \Rubix\ML\Graph\Nodes\BinaryNode $node
-     */
-    public function attachLeft(?BinaryNode $node = null) : void
-    {
-        $this->left = $node;
-    }
-
-    /**
-     * Set the right child node.
-     *
-     * @param \Rubix\ML\Graph\Nodes\BinaryNode $node
-     */
-    public function attachRight(?BinaryNode $node = null) : void
-    {
-        $this->right = $node;
     }
 
     /**
